fetch realtime for tel aviv (test)

+++ lib/features/realtime_weather/domain/entities/realtime_weather.dart
@@ -2,8 +2,8 @@ import 'package:equatable/equatable.dart';
 
 class RealtimeWeatherEntity extends Equatable {
   final String? loactionName;
-  final int? tempC;
-  final int? tempF;
+  final double? tempC;
+  final double? tempF;
   final int? isDay;
   final String? conditionText;
   final String? conditionIcon;
